feat: add temp way to deploy
At the moment we've got no way to deploy other than manually running
this crap on our machines.

That isn't as 'optimal' as it should be so lets for the time being run
this on GitHub Actions.

There's a crap load of improvements and reasons why doing it like this
is horrible but who cares.

yolo swaggings 420 blaze it.

* feat: add kim as a user so it can deploy
* feat: add dry run on pull_request
* feat: add secrets and fmt
* feat: sshUser as kim
* feat: do some terrible things